requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
Django==3.0.8
django-braces==1.14.0
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-tastypie==0.14.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
oauthlib==3.1.0
packaging==20.4
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-mimeparse==1.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
mysqlclient==2.0.1